--- a/lib/app/client.js
+++ b/lib/app/client.js
@@ -2,12 +2,14 @@
 
 import Vue from 'vue'
 import middleware from './middleware'
-import { app, router<%= (store ? ', store' : '') %>, NuxtError } from './index'
+import { createApp, NuxtError } from './index'
 import { applyAsyncData, getMatchedComponents, getMatchedComponentsInstances, flatMapComponents, getContext, promiseSeries, promisify, getLocation, compile } from './utils'
 const noopData = () => { return {} }
 const noopFetch = () => {}
 let _lastPaths = []
 let _lastComponentsFiles = []
+
+const { app, router<%= (store ? ', store' : '') %> } = createApp()
 
 function mapTransitions(Components, to, from) {
   return Components.map((Component) => {