--- a/lib/app/router.js
+++ b/lib/app/router.js
@@ -53,12 +53,14 @@ const scrollBehavior = (to, from, savedPosition) => {
 }
 <% } %>
 
-export default new Router({
-  mode: '<%= router.mode %>',
-  base: '<%= router.base %>',
-  linkActiveClass: '<%= router.linkActiveClass %>',
-  scrollBehavior,
-  routes: [
-<%= _routes %>
-  ]
-})
+export function createRouter () {
+  return new Router({
+    mode: '<%= router.mode %>',
+    base: '<%= router.base %>',
+    linkActiveClass: '<%= router.linkActiveClass %>',
+    scrollBehavior,
+    routes: [
+  <%= _routes %>
+    ]
+  })
+}