--- a/lib/app/server.js
+++ b/lib/app/server.js
@@ -7,11 +7,10 @@ import Vue from 'vue'
 import { stringify } from 'querystring'
 import { omit } from 'lodash'
 import middleware from './middleware'
-import { app, router<%= (store ? ', store' : '') %>, NuxtError } from './index'
+import { createApp, NuxtError } from './index'
 import { applyAsyncData, getMatchedComponents, getContext, promiseSeries, promisify, urlJoin } from './utils'
 
 const isDev = <%= isDev %>
-const _app = new Vue(app)
 
 // This exported function will be called by `bundleRenderer`.
 // This is where we perform data-prefetching to determine the
@@ -19,6 +18,8 @@ const _app = new Vue(app)
 // Since data fetching is async, this function is expected to
 // return a Promise that resolves to the app instance.
 export default context => {
+  const { app, router<%= (store ? ', store' : '') %> } = createApp(context)
+  const _app = new Vue(app)
   // Add store to the context
   <%= (store ? 'context.store = store' : '') %>
   // Nuxt object