--- a/lib/app/store.js
+++ b/lib/app/store.js
@@ -30,8 +30,11 @@ let storeData = {}
 if (filenames.indexOf('./index.js') !== -1) {
   let mainModule = getModule('./index.js')
   if (mainModule.commit) {
-    store = mainModule
+    console.error('[nuxt.js] store/index.js should export raw store options instead of an instance.')
   } else {
+    if (mainModule.state && typeof mainModule.state !== 'function') {
+      console.error('[nuxt.js] store state should be a function.')
+    }
     storeData = mainModule
   }
 }
@@ -49,8 +52,13 @@ if (store == null) {
     name = namePath.pop()
     module[name] = getModule(filename)
     module[name].namespaced = true
+
+    if (typeof module[name].state !== 'function') {
+      console.error('[nuxt.js] store module state should be a function.')
+    }
   }
-  store = new Vuex.Store(storeData)
 }
 
-export default store
+export function createStore () {
+  return new Vuex.Store(storeData)
+}