--- a/lib/build.js
+++ b/lib/build.js
@@ -464,7 +464,8 @@ function createRenderer (bundle) {
     }))
   }
   this.renderer = createBundleRenderer(bundle, {
-    cache: cacheConfig
+    cache: cacheConfig,
+    runInNewContext: false
   })
   this.renderToString = pify(this.renderer.renderToString)
   this.renderToStream = this.renderer.renderToStream