This PR contains changes necessary to make use of the new `runInNewContext: false` option introduced in 2.3.

The changes are implemented and tested on top of https://github.com/nuxt/nuxt.js/pull/633 and should be merged after it. (This is probably why CI is broken, this is mainly submitted to show the diffs, I can re-submit after #633 is merged as well)

**This introduces breaking changes.** Most notably the user will have to avoid global singletons in their code as explained [here](https://ssr.vuejs.org/en/structure.html). Also, in [store.js](https://github.com/yyx990803/nuxt.js/blob/context/lib/app/store.js#L33) we have to add a few warnings to require user store and modules to use functions for `state`.

It's possible to maintain backwards compat by making this whole thing behind a config option, but I think this really should be the new default.

Perf comparison for an out-of-the-box Nuxt app using `vue init nuxt/starter` with `ab -n 1000`:

### Before

```
Concurrency Level:      1
Time taken for tests:   10.904 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      4531000 bytes
HTML transferred:       4371000 bytes
Requests per second:    91.71 [#/sec] (mean)
Time per request:       10.904 [ms] (mean)
Time per request:       10.904 [ms] (mean, across all concurrent requests)
Transfer rate:          405.81 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:     6   11   6.5      9      92
Waiting:        6   11   6.4      9      91
Total:          6   11   6.5      9      92

Percentage of the requests served within a certain time (ms)
  50%      9
  66%     10
  75%     10
  80%     11
  90%     13
  95%     23
  98%     32
  99%     35
 100%     92 (longest request)
```

### After

```
Concurrency Level:      1
Time taken for tests:   2.126 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      4531000 bytes
HTML transferred:       4371000 bytes
Requests per second:    470.41 [#/sec] (mean)
Time per request:       2.126 [ms] (mean)
Time per request:       2.126 [ms] (mean, across all concurrent requests)
Transfer rate:          2081.49 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:     1    2   1.8      1      17
Waiting:        1    2   1.8      1      17
Total:          1    2   1.8      1      17

Percentage of the requests served within a certain time (ms)
  50%      1
  66%      2
  75%      2
  80%      2
  90%      3
  95%      5
  98%      9
  99%     13
 100%     17 (longest request)
```