@pi0 2017-05-02
I've successfully cherry-picked commits into #633 and CI is now passing tests :)  @yyx990803